tidy up and avoid duplication
